+++ src/renderer/components/provider-details/index.ts
@@ -1,3 +1,0 @@
-export * from './traefik'
-export * from './istio'
-export * from './nginx'